% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getDataPSQL}
\alias{getDataPSQL}
\title{Run Sample Query via \code{psql}}
\usage{
getDataPSQL(sqlquery)
}
\arguments{
\item{sqlquery}{A SQL query as character string}
}
\value{
The dataset corresponding to the query.
}
\description{
Run a sample query against Vertica using \code{psql}.
}
\details{
The \code{psql} binary is required.
}
\author{
Dirk Eddelbuettel
}

