% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getDataPython}
\alias{getDataPython}
\title{Run Sample Query via Python}
\usage{
getDataPython(sqlquery)
}
\arguments{
\item{sqlquery}{A SQL query as character string}
}
\value{
The dataset corresponding to the query.
}
\description{
Run a sample query against Vertica using the Python-Vertica
package.
}
\details{
The Python installation must have the Python-Vertica package
contributed by Uber. There is essentially no error handling. This
function is really only here for completeness and not recommended.
The rPython package could be used as well.
}
\author{
Dirk Eddelbuettel
}

