% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getDataRPostgreSQL}
\alias{getDataRPostgreSQL}
\title{Run Sample Query via RPostgreSQL}
\usage{
getDataRPostgreSQL(sqlquery)
}
\arguments{
\item{sqlquery}{A SQL query as character string}
}
\value{
The dataset corresponding to the query.
}
\description{
Run a sample query against Vertica using RPostgreSQL.
}
\details{
The RPostgreSQL package is used, and has to be configured via
environment variables.
}
\author{
Dirk Eddelbuettel
}

