% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{psqlCommand}
\alias{psqlCommand}
\title{Run Simple SQL Command via \code{psql}}
\usage{
psqlCommand(sqlcmd)
}
\arguments{
\item{sqlcmd}{A SQL cmd as character string}
}
\value{
The result of running the command.
}
\description{
Simple convenience wrapper to run a single command
}
\details{
The \code{psql} binary is required.
}
\author{
Dirk Eddelbuettel
}

