% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{psqlCopy}
\alias{psqlCopy}
\title{Write Data to a Table}
\usage{
psqlCopy(data, table)
}
\arguments{
\item{data}{A matrix or data.frame which is to be written}

\item{table}{The name of existing table}
}
\value{
Nothing
}
\description{
Write a data object to a table using \code{psql}
}
\details{
Data is written via pipe into \code{psql} which reads the data and
inserts it into the give table.  There is no error checking. Proof
of concept.
}
\author{
Dirk Eddelbuettel
}

