% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{runBenchmarks}
\alias{runBenchmarks}
\title{Run Simple Vertica Benchmark Query}
\usage{
runBenchmarks(query = "SELECT * FROM product_dimension LIMIT 10000;")
}
\arguments{
\item{query}{A SQL query, default uses the example data set.}
}
\value{
A data.frame with the first for columns returned by
 \code{benchmark}.
}
\description{
Run benchmarks against Vertica using three different access methods:
the \code{psql} binary, the RJDBC package and the RPostgreSQL package.
}
\author{
Dirk Eddelbuettel
}

